+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog track changes to the qoqo project starting at version 0.1.0
 
 ## Not released
 
+### Added not release
+
+* CMake based cmake-find-openmp helper crate to correctly link openmp
+
 ## 0.2.0
 
 ### Fixed 0.2.0